General Improvements
Now assigning to 2-d matrix is automatically done in the first for loop which saves quite a bit of time for the large matrix sizes.  Also put some lables. There is an offset on the energy term that I put for the time being without a motivation, just to try and shift the energy axis to zero. Need to look back on this since I haven't thoght too much about it.