Things to change
NOTE: NOT A STABLE UPDATE
What happened is, about a month ago I tested the function with a simple square lattice. The Wannier functions are good. The next thing to check was the U and J values. However, I noticed that these changed with the variable numbands, which it shouldn't! Since this decides the number of bands to compute beforehand,  but as long as there are enough, only the variable bands = [1] should deicde which bands are actually included in the calculation of the final wannier functions.

Note that all of the recent changes have been done on numerics testing, and once things are looking good there I will port the changes over to lattice wannier