Better Reciprocal Lattice Handling
This is a relatively stable update. From the last time, I switched from the bad way I was doing things before to a more solid-state oriented way using an actual reciprocal lattice vector basis. This made the magnitudes of the hopping coefficients for the square lattice(which I am testing right now to get things right) actually match in rough magnitude.

Now it's time to get the hopping parameter vs lattice depth looking better, since right now the graph produced with meta code looks terrible, and I think that this is due to the finite size effects of the wannier code.