Full BH Parameters Calculation and Plotting
This version includes a lot of updates to the structure of the code, where I tried to move functionality to outside functions.

Perhaps the most exciting part is that meta_code outputs a full BH parameters plot that scales with lattice depth qualitatively right. The only thing is that there is a constant factor, so I need to be a little more careful in the calcuations; I think I already know what it is for U, but for J I need to double check, since for now I can't see quite what it is. One thing that is very suspect right now is normalization.